+++ src/components/scenes/MainMenu.js
@@ -2,18 +2,36 @@ import { Scene } from "../../modules/scenemanagement/scene.js";
 import { UI } from "../ui/ui.js";
 import { global } from "../../modules/global.js";
 import { Button } from "../ui/button.js";
+import { GameWorld } from "./GameWorld.js";
 
 export class MainMenu extends Scene {
+
+    gap = 12;
+    theme = null;
 
     constructor() {
         super("Main Menu");
         this.sceneName = "Main Menu";
         this.sceneObjects = [];
-        this.start = new Button(global.canvas.width / 2, global.canvas.height / 2, 200, 50, "Start", 20, "black", "beige", () => {  
-            //let theme = new Audio('src/components/audio/TwelveMonths - Hundred Cups of Tin.mp3', false, 0.5);
-            //theme.play();
-            console.log("yems");
+        this.start = new Button((global.canvas.width / 2) - 128, global.canvas.height / 2 - 128, 256, 64, "Start", 20, "black", "beige", () => {  
+            // let theme = new Audio("../tmhcot.mp3", false, 0.5);
+            // theme.play();
+            global.sceneManager.changeScene(new GameWorld());
+            console.log("start");
         });
+        this.options = new Button((global.canvas.width / 2) - 128, (global.canvas.height / 2) - 64 + this.gap, 256, 64, "Options", 20, "black", "beige", () => {
+            console.log("options");
+        });
+        this.quit = new Button((global.canvas.width / 2) - 128, (global.canvas.height / 2) + (this.gap * 2), 256, 64, "Quit", 20, "black", "beige", () => {
+            console.log("quit");
+        });
+
+        this.theme = new Audio('/src/components/audio/TwelveMonths - Hundred Cups of Tin.ogg', true, 0.5);
+
+        this.sceneObjects.push(this.start);
+        this.sceneObjects.push(this.options);
+        this.sceneObjects.push(this.quit);
+
         console.log(`Scene ${this.sceneName} constructed`);
 
     }
@@ -21,8 +39,10 @@ export class MainMenu extends Scene {
     init = function() {
         this.createObjects();
 
-        let theme = new Audio('src/components/audio/TwelveMonths - Hundred Cups of Tin.mp3', false, 0.5);
-        theme.play();
+        this.theme.play();
+
+        console.log(this.theme);
+
 
         console.log(`Scene ${this.sceneName} initialized`);
     }
@@ -31,37 +51,35 @@ export class MainMenu extends Scene {
 
     }
 
-    render = function() {
+    render = function(ctx) {
         for(let i = 0; i < this.sceneObjects.length; i++) {
-            this.sceneObjects[i].render();
+            this.sceneObjects[i].render(ctx);
         }
     }
 
     destroy = function() {
-        // if (this.destroyObjects()) {
-        //     console.log(`Scene ${this.sceneName} destroyed`);
-        // }
+        //this.theme.stop();
+        //destroyObjects();
     }
 
     createObjects = function() {
 
-
-        this.sceneObjects.push(this.start);
     }
 
-    destroyObjects = function(sceneObjects) {
+    destroyObjects() {
         // Check if there are objects in the scene
         if (this.sceneObjects.length > 0) {
-            // Destroy each object
+        // Destroy each object
             this.sceneObjects.forEach(object => {
-                object.destroy();
+                if(object.destroy)
+                {
+                    object.destroy();
+                }
             });
 
             // Clear the sceneObjects array
             this.sceneObjects = [];
             theme.stop();
-
-            return true;
         }
     }
 }