+++ src/main.js
@@ -12,6 +12,9 @@ function gameLoop(totalRunningTime) {
     // STATS
     global.dt(totalRunningTime);
     global.updateFPS();
+
+    global.ctx.font = "16px Pixelify Sans";
+    global.ctx.fillStyle = "white";
     global.ctx.fillText(`FPS: ${Math.round(global.fps)} - ∂t: ${global.deltaTime}`, global.getCanvasBounds().left+64, global.getCanvasBounds().bottom-80);
     global.ctx.fillText(`left: ${global.getCanvasBounds().left}\ntop: ${global.getCanvasBounds().top}\nright: ${global.getCanvasBounds().right}\nbottom: ${global.getCanvasBounds().bottom}\n`, global.getCanvasBounds().left+64, global.getCanvasBounds().bottom-64);